+++ sql/src/main/java/org/apache/druid/sql/http/SqlResource.java
@@ -19,27 +19,19 @@
 
 package org.apache.druid.sql.http;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.sun.jersey.api.core.HttpContext;
-import org.apache.druid.common.exception.SanitizableException;
 import org.apache.druid.error.DruidException;
-import org.apache.druid.guice.annotations.Self;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.query.DefaultQueryConfig;
 import org.apache.druid.query.QueryContext;
 import org.apache.druid.query.QueryContexts;
-import org.apache.druid.server.DruidNode;
-import org.apache.druid.server.QueryLifecycle;
 import org.apache.druid.server.QueryResource;
-import org.apache.druid.server.QueryResponse;
 import org.apache.druid.server.QueryResultPusher;
-import org.apache.druid.server.ResponseContextConfig;
-import org.apache.druid.server.initialization.ServerConfig;
 import org.apache.druid.server.security.Action;
 import org.apache.druid.server.security.AuthenticationResult;
 import org.apache.druid.server.security.AuthorizationResult;
@@ -47,12 +39,10 @@ import org.apache.druid.server.security.AuthorizationUtils;
 import org.apache.druid.server.security.AuthorizerMapper;
 import org.apache.druid.server.security.Resource;
 import org.apache.druid.server.security.ResourceAction;
-import org.apache.druid.sql.DirectStatement.ResultSet;
 import org.apache.druid.sql.HttpStatement;
 import org.apache.druid.sql.SqlLifecycleManager;
 import org.apache.druid.sql.SqlLifecycleManager.Cancelable;
 import org.apache.druid.sql.SqlQueryPlus;
-import org.apache.druid.sql.SqlRowTransformer;
 import org.apache.druid.sql.calcite.run.SqlEngine;
 
 import javax.annotation.Nullable;
@@ -68,12 +58,9 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -88,38 +75,30 @@ public class SqlResource
   public static final String SQL_HEADER_VALUE = "yes";
 
   private static final Logger log = new Logger(SqlResource.class);
-  private static final SqlResourceQueryMetricCounter QUERY_METRIC_COUNTER = new SqlResourceQueryMetricCounter();
+  public static final SqlResourceQueryMetricCounter QUERY_METRIC_COUNTER = new SqlResourceQueryMetricCounter();
 
-  private final ObjectMapper jsonMapper;
   private final AuthorizerMapper authorizerMapper;
-  private final ServerConfig serverConfig;
-  private final ResponseContextConfig responseContextConfig;
-  private final DefaultQueryConfig defaultQueryConfig;
-  private final DruidNode selfNode;
+  private final SqlResourceQueryResultPusherFactory resultPusherFactory;
   private final SqlLifecycleManager sqlLifecycleManager;
   private final SqlEngineRegistry sqlEngineRegistry;
+  private final DefaultQueryConfig defaultQueryConfig;
 
   @VisibleForTesting
   @Inject
   public SqlResource(
-      final ObjectMapper jsonMapper,
       final AuthorizerMapper authorizerMapper,
-      final ServerConfig serverConfig,
       final SqlLifecycleManager sqlLifecycleManager,
       final SqlEngineRegistry sqlEngineRegistry,
-      final ResponseContextConfig responseContextConfig,
-      final DefaultQueryConfig defaultQueryConfig,
-      @Self final DruidNode selfNode
+      final SqlResourceQueryResultPusherFactory resultPusherFactory,
+      final DefaultQueryConfig defaultQueryConfig
   )
   {
+    this.resultPusherFactory = resultPusherFactory;
     this.sqlEngineRegistry = Preconditions.checkNotNull(sqlEngineRegistry, "sqlEngineRegistry");
-    this.jsonMapper = Preconditions.checkNotNull(jsonMapper, "jsonMapper");
     this.authorizerMapper = Preconditions.checkNotNull(authorizerMapper, "authorizerMapper");
-    this.serverConfig = Preconditions.checkNotNull(serverConfig, "serverConfig");
-    this.responseContextConfig = responseContextConfig;
-    this.defaultQueryConfig = Preconditions.checkNotNull(defaultQueryConfig, "defaultQueryConfig");
-    this.selfNode = selfNode;
     this.sqlLifecycleManager = Preconditions.checkNotNull(sqlLifecycleManager, "sqlLifecycleManager");
+    this.defaultQueryConfig = Preconditions.checkNotNull(defaultQueryConfig, "defaultQueryConfig");
+
   }
 
   @GET
@@ -207,7 +186,9 @@ public class SqlResource
     final String currThreadName = Thread.currentThread().getName();
     try {
       Thread.currentThread().setName(StringUtils.format("sql[%s]", stmt.sqlQueryId()));
-      return makePusher(req, stmt, sqlQuery, queryContext).push();
+
+      QueryResultPusher pusher = resultPusherFactory.factorize(req, stmt, sqlQuery);
+      return pusher.push();
     }
     finally {
       Thread.currentThread().setName(currThreadName);
@@ -265,169 +246,6 @@ public class SqlResource
     @Override
     public void incrementTimedOut()
     {
-    }
-  }
-
-  private SqlResourceQueryResultPusher makePusher(
-      HttpServletRequest req,
-      HttpStatement stmt,
-      SqlQuery sqlQuery,
-      QueryContext queryContext
-  )
-  {
-    final String sqlQueryId = stmt.sqlQueryId();
-    Map<String, String> headers = new LinkedHashMap<>();
-    headers.put(SQL_QUERY_ID_RESPONSE_HEADER, sqlQueryId);
-
-    if (sqlQuery.includeHeader()) {
-      headers.put(SQL_HEADER_RESPONSE_HEADER, SQL_HEADER_VALUE);
-    }
-
-    return new SqlResourceQueryResultPusher(req, sqlQueryId, stmt, sqlQuery, queryContext, headers);
-  }
-
-  private class SqlResourceQueryResultPusher extends QueryResultPusher
-  {
-    private final String sqlQueryId;
-    private final HttpStatement stmt;
-    private final SqlQuery sqlQuery;
-
-    /**
-     * Context to use for pushing results. May be different from the context in SqlQuery due to SET statements.
-     */
-    private final QueryContext queryContext;
-
-    public SqlResourceQueryResultPusher(
-        HttpServletRequest req,
-        String sqlQueryId,
-        HttpStatement stmt,
-        SqlQuery sqlQuery,
-        QueryContext queryContext,
-        Map<String, String> headers
-    )
-    {
-      super(
-          req,
-          jsonMapper,
-          responseContextConfig,
-          selfNode,
-          SqlResource.QUERY_METRIC_COUNTER,
-          sqlQueryId,
-          MediaType.APPLICATION_JSON_TYPE,
-          headers
-      );
-      this.sqlQueryId = sqlQueryId;
-      this.stmt = stmt;
-      this.queryContext = queryContext;
-      this.sqlQuery = sqlQuery;
-    }
-
-    @Override
-    public ResultsWriter start()
-    {
-      return new ResultsWriter()
-      {
-        private QueryResponse<Object[]> queryResponse;
-        private ResultSet thePlan;
-
-        @Override
-        @Nullable
-        public Response.ResponseBuilder start()
-        {
-          thePlan = stmt.plan();
-          queryResponse = thePlan.run();
-          return null;
-        }
-
-        @Override
-        @SuppressWarnings({"unchecked", "rawtypes"})
-        public QueryResponse<Object> getQueryResponse()
-        {
-          return (QueryResponse) queryResponse;
-        }
-
-        @Override
-        public Writer makeWriter(OutputStream out) throws IOException
-        {
-          ResultFormat.Writer writer = sqlQuery.getResultFormat().createFormatter(out, jsonMapper);
-          final SqlRowTransformer rowTransformer = thePlan.createRowTransformer();
-
-          return new Writer()
-          {
-
-            @Override
-            public void writeResponseStart() throws IOException
-            {
-              writer.writeResponseStart();
-
-              if (sqlQuery.includeHeader()) {
-                writer.writeHeader(
-                    rowTransformer.getRowType(),
-                    sqlQuery.includeTypesHeader(),
-                    sqlQuery.includeSqlTypesHeader()
-                );
-              }
-            }
-
-            @Override
-            public void writeRow(Object obj) throws IOException
-            {
-              Object[] row = (Object[]) obj;
-
-              writer.writeRowStart();
-              for (int i = 0; i < rowTransformer.getFieldList().size(); i++) {
-                final Object value = rowTransformer.transform(row, i);
-                writer.writeRowField(rowTransformer.getFieldList().get(i), value);
-              }
-              writer.writeRowEnd();
-            }
-
-            @Override
-            public void writeResponseEnd() throws IOException
-            {
-              writer.writeResponseEnd();
-            }
-
-            @Override
-            public void close() throws IOException
-            {
-              writer.close();
-            }
-          };
-        }
-
-        @Override
-        public void recordSuccess(long numBytes)
-        {
-          stmt.reporter().succeeded(numBytes);
-        }
-
-        @Override
-        public void recordFailure(Exception e)
-        {
-          if (QueryLifecycle.shouldLogStackTrace(e, queryContext)) {
-            log.warn(e, "Exception while processing sqlQueryId[%s]", sqlQueryId);
-          } else {
-            log.noStackTrace().warn(e, "Exception while processing sqlQueryId[%s]", sqlQueryId);
-          }
-          stmt.reporter().failed(e);
-        }
-
-        @Override
-        public void close()
-        {
-          stmt.close();
-        }
-      };
-    }
-
-    @Override
-    public void writeException(Exception ex, OutputStream out) throws IOException
-    {
-      if (ex instanceof SanitizableException) {
-        ex = serverConfig.getErrorResponseTransformStrategy().transformIfNeeded((SanitizableException) ex);
-      }
-      out.write(jsonMapper.writeValueAsBytes(ex));
     }
   }
